@@ -56,40 +92,49 @@
       "last 1 safari version"
     ]
   },
+  "eslintConfig": {
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
+  },
+  "dependencies": {
+    "@rsksmart/rif-relay-client": "github:rsksmart/rif-relay-client#PP-357/dapp-update",
+    "@rsksmart/rif-relay-contracts": "github:rsksmart/rif-relay-contracts#feature/PP-278/PP-329/upgrade_contracts",
+    "@rsksmart/rlogin": "^1.4.1-beta.1",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^13.5.0",
+    "@types/jest": "^27.5.2",
+    "@types/node": "^16.18.10",
+    "@types/react": "^17.0.39",
+    "@types/react-dom": "^18.0.10",
+    "bignumber.js": "^9.1.1",
+    "ethers": "^5.7.2",
+    "materialize-css": "^1.0.0",
+    "qrcode.react": "^3.1.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-materialize": "^3.10.0",
+    "react-scripts": "5.0.1",
+    "typescript": "4.8.2",
+    "web-vitals": "^2.1.4"
+  },
   "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^5.13.0",
-    "@typescript-eslint/parser": "^5.13.0",
-    "dotenv": "^10.0.0",
-    "eslint": "^7.32.0",
+    "@commitlint/cli": "^17.0.3",
+    "@commitlint/config-conventional": "^17.0.3",
+    "@tsconfig/node16-strictest": "^1.0.4",
+    "@typescript-eslint/eslint-plugin": "^5.47.0",
+    "@typescript-eslint/parser": "^5.47.0",
+    "eslint": "^8.31.0",
     "eslint-config-airbnb": "^19.0.4",
-    "eslint-config-airbnb-typescript": "^16.1.0",
+    "eslint-config-airbnb-typescript": "^17.0.0",
     "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-import": "^2.25.4",
-    "eslint-plugin-jsx-a11y": "^6.5.1",
-    "eslint-plugin-react": "^7.29.3",
-    "eslint-plugin-react-hooks": "^4.3.0",
-    "husky": "^7.0.4",
-    "keyv": "^4.5.2",
-    "lint-staged": "^12.3.5",
-    "prettier": "^2.5.1"
-  },
-  "description": "This project was bootstrapped with [Create React App](https://github.com/facebook/create-react-app).",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/@rsksmart/rif-relay-sample-dapp"
-  },
-  "keywords": [
-    "SDK",
-    "relay",
-    "services"
-  ],
-  "author": "Wilson Hamilton",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/@rsksmart/rif-relay-sample-dapp/issues"
-  },
-  "lint-staged": {
-    "*.{ts,tsx}": "eslint --cache --fix",
-    "*.{js,css,md}": "prettier --write"
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-jsx-a11y": "^6.6.1",
+    "eslint-plugin-react": "^7.31.11",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "husky": "^8.0.1",
+    "prettier": "^2.7.1"
   }
 }
