@@ -1,5 +1,5 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
+
+import ReactDOM from 'react-dom/client';
 import 'src/index.css';
 import App from 'src/App';
 import reportWebVitals from 'src/reportWebVitals';
@@ -7,13 +7,15 @@ import { StoreProvider } from './context/context';
 import 'materialize-css/dist/css/materialize.min.css';
 import 'materialize-css/dist/js/materialize.min';
 
-ReactDOM.render(
-    <React.StrictMode>
+const root = ReactDOM.createRoot(
+    document.getElementById('root') as HTMLElement
+);
+root.render(
+   // <React.StrictMode>
         <StoreProvider>
             <App />
         </StoreProvider>
-    </React.StrictMode>,
-    document.getElementById('root')
+    //</React.StrictMode>
 );
 
 // If you want to start measuring performance in your app, pass a function
