@@ -1,6 +1,7 @@
-import { RelayingServices, ERC20Token } from '@rsksmart/rif-relay-sdk';
-import { ReactNode } from 'react';
-import { Modals, SmartWalletWithBalance } from 'src/types';
+import type { RelayClient } from '@rsksmart/rif-relay-client';
+import type { providers } from 'ethers';
+import type { ReactNode } from 'react';
+import type { ERC20Token, Modals, SmartWallet } from 'src/types';
 
 export const RELOAD_ACTION = 'reload';
 export const RELOAD_PARTNERS_ACTION = 'reload_partners';
@@ -16,53 +17,59 @@ export const SET_SMART_WALLETS_ACTION = 'set_smart_wallets';
 export const ADD_SMART_WALLET_ACTION = 'add_smart_wallet';
 export const UPDATE_SMART_WALLET_ACTION = 'update_smart_wallet';
 export const SET_MODALS_ACTION = 'set_modals';
+export const SET_RELAY_CLIENT_ACTION = 'set_relay_client';
 
 export type Action =
-    | { type: typeof RELOAD_ACTION; reload: boolean }
-    | { type: typeof RELOAD_PARTNERS_ACTION; reloadPartners: boolean }
-    | { type: typeof RELOAD_TOKEN_ACTION; reloadToken: boolean }
-    | { type: typeof SET_ACCOUNT_ACTION; account: string }
-    | { type: typeof SET_CONNECTED_ACTION; connected: boolean }
-    | { type: typeof SET_PROVIDER_ACTION; provider: RelayingServices }
-    | { type: typeof SET_CHAIN_ID_ACTION; chainId: number }
-    | { type: typeof SET_LOADER_ACTION; loader: boolean }
-    | { type: typeof SET_TOKEN_ACTION; token: ERC20Token | undefined }
-    | {
-          type: typeof SET_SMART_WALLET_ACTION;
-          smartWallet: SmartWalletWithBalance;
-      }
-    | {
-          type: typeof SET_SMART_WALLETS_ACTION;
-          smartWallets: SmartWalletWithBalance[];
-      }
-    | {
-          type: typeof ADD_SMART_WALLET_ACTION;
-          smartWallet: SmartWalletWithBalance;
-      }
-    | {
-          type: typeof UPDATE_SMART_WALLET_ACTION;
-          smartWallet: SmartWalletWithBalance;
-      }
-    | {
-          type: typeof SET_MODALS_ACTION;
-          modal: Partial<Modals>;
-      };
+  | { type: typeof RELOAD_ACTION; reload: boolean }
+  | { type: typeof RELOAD_PARTNERS_ACTION; reloadPartners: boolean }
+  | { type: typeof RELOAD_TOKEN_ACTION; reloadToken: boolean }
+  | { type: typeof SET_ACCOUNT_ACTION; account: string }
+  | { type: typeof SET_CONNECTED_ACTION; connected: boolean }
+  | { type: typeof SET_PROVIDER_ACTION; provider: providers.JsonRpcProvider }
+  | { type: typeof SET_CHAIN_ID_ACTION; chainId: number }
+  | { type: typeof SET_LOADER_ACTION; loader: boolean }
+  | { type: typeof SET_TOKEN_ACTION; token: ERC20Token | undefined }
+  | {
+    type: typeof SET_SMART_WALLET_ACTION;
+    smartWallet: SmartWallet;
+  }
+  | {
+    type: typeof SET_SMART_WALLETS_ACTION;
+    smartWallets: SmartWallet[];
+  }
+  | {
+    type: typeof ADD_SMART_WALLET_ACTION;
+    smartWallet: SmartWallet;
+  }
+  | {
+    type: typeof UPDATE_SMART_WALLET_ACTION;
+    smartWallet: SmartWallet;
+  }
+  | {
+    type: typeof SET_MODALS_ACTION;
+    modal: Partial<Modals>;
+  }
+  | {
+    type: typeof SET_RELAY_CLIENT_ACTION;
+    relayClient: RelayClient;
+  };
 
 export type Dispatch = (action: Action) => void;
 
 export type ProviderProps = { children: ReactNode };
 
 export type State = {
-    account: string;
-    connected: boolean;
-    provider: RelayingServices | undefined;
-    chainId: number;
-    loader: boolean;
-    token: ERC20Token | undefined;
-    smartWallet: SmartWalletWithBalance | undefined;
-    reload: boolean;
-    reloadPartners: boolean;
-    reloadToken: boolean;
-    modals: Modals;
-    smartWallets: SmartWalletWithBalance[];
+  account: string;
+  connected: boolean;
+  provider: providers.JsonRpcProvider | undefined;
+  chainId: number;
+  loader: boolean;
+  token: ERC20Token | undefined;
+  smartWallet: SmartWallet | undefined;
+  reload: boolean;
+  reloadPartners: boolean;
+  reloadToken: boolean;
+  modals: Modals;
+  smartWallets: SmartWallet[];
+  relayClient: RelayClient | undefined;
 };
